Moved usb gps code to separate subfolder
