+++ src/CMakeLists.txt
@@ -178,7 +178,7 @@ add_library(real_world_interface STATIC
   ar/MarkerPattern.cpp
   ar/Tag.cpp
   ar/read_landmarks.cpp
-  gps/read_usb_gps.cpp
+  gps/usb_gps/read_usb_gps.cpp
   )
 target_include_directories(real_world_interface SYSTEM PUBLIC ${OpenCV_INCLUDE_DIRS})
 target_link_libraries(real_world_interface ${vision_libs})
